+++ og_marl/tf2_systems/offline/iddpg_bc.py
@@ -1,0 +1,325 @@
+# Copyright 2023 InstaDeep Ltd. All rights reserved.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Implementation of IDDPG+CQL"""
+from typing import Any, Dict, Optional, Tuple
+
+import copy
+import hydra
+import numpy as np
+from omegaconf import DictConfig
+import tensorflow as tf
+from tensorflow import Tensor
+import sonnet as snt
+import tree
+from chex import Numeric
+
+from og_marl.environments import get_environment, BaseEnvironment
+from og_marl.loggers import BaseLogger, WandbLogger
+from og_marl.vault_utils.download_vault import download_and_unzip_vault
+from og_marl.replay_buffers import Experience, FlashbaxReplayBuffer
+from og_marl.tf2_systems.networks import StateAndActionCritic
+from og_marl.tf2_systems.offline.base import BaseOfflineSystem
+from og_marl.tf2_systems.utils import (
+    batch_concat_agent_id_to_obs,
+    concat_agent_id_to_obs,
+    expand_batch_and_agent_dim_of_time_major_sequence,
+    merge_batch_and_agent_dim_of_time_major_sequence,
+    set_growing_gpu_memory,
+    switch_two_leading_dims,
+    unroll_rnn,
+)
+
+set_growing_gpu_memory()
+
+
+class IDDPGBCSystem(BaseOfflineSystem):
+    """Independent DDPG + BC.
+
+    NOTE: the critic conditions on states and individual agent actions.
+
+    """
+
+    def __init__(
+        self,
+        environment: BaseEnvironment,
+        logger: BaseLogger,
+        linear_layer_dim: int = 64,
+        recurrent_layer_dim: int = 64,
+        discount: float = 0.99,
+        target_update_rate: float = 0.005,
+        critic_learning_rate: float = 3e-4,
+        policy_learning_rate: float = 3e-4,
+        add_agent_id_to_obs: bool = True,
+        bc_alpha: float = 2.5,  # BC
+    ):
+        super().__init__(
+            environment=environment,
+            logger=logger,
+        )
+
+        self.discount = discount
+        self.add_agent_id_to_obs = add_agent_id_to_obs
+
+        # Policy network
+        self.policy_network = snt.DeepRNN(
+            [
+                snt.Linear(linear_layer_dim),
+                tf.nn.relu,
+                snt.GRU(recurrent_layer_dim),
+                tf.nn.relu,
+                snt.Linear(self.environment.num_actions),
+                tf.nn.tanh,
+            ]
+        )  # shared network for all agents
+
+        # Target policy network
+        self.target_policy_network = copy.deepcopy(self.policy_network)
+
+        # Critic network
+        self.critic_network_1 = StateAndActionCritic(
+            len(self.environment.agents), self.environment.num_actions, add_agent_id_to_obs
+        )  # shared network for all agents
+        self.critic_network_2 = copy.deepcopy(self.critic_network_1)
+
+        # Target critic network
+        self.target_critic_network_1 = copy.deepcopy(self.critic_network_1)
+        self.target_critic_network_2 = copy.deepcopy(self.critic_network_1)
+        self.target_update_rate = target_update_rate
+
+        # Optimizers
+        self.critic_optimizer = snt.optimizers.Adam(learning_rate=critic_learning_rate)
+        self.policy_optimizer = snt.optimizers.Adam(learning_rate=policy_learning_rate)
+
+        # Reset the recurrent neural network
+        self.rnn_states = {
+            agent: self.policy_network.initial_state(1) for agent in self.environment.agents
+        }
+
+        self.bc_alpha = bc_alpha
+
+    def reset(self) -> None:
+        """Called at the start of a new episode."""
+        # Reset the recurrent neural network
+        self.rnn_states = {
+            agent: self.policy_network.initial_state(1) for agent in self.environment.agents
+        }
+        return
+
+    def select_actions(
+        self,
+        observations: Dict[str, np.ndarray],
+        legal_actions: Optional[Dict[str, np.ndarray]] = None,
+    ) -> Dict[str, np.ndarray]:
+        actions, next_rnn_states = self._tf_select_actions(observations, self.rnn_states)
+        self.rnn_states = next_rnn_states
+        return tree.map_structure(  # type: ignore
+            lambda x: x[0].numpy(), actions
+        )  # convert to numpy and squeeze batch dim
+
+    @tf.function(jit_compile=True)
+    def _tf_select_actions(
+        self,
+        observations: Dict[str, Tensor],
+        rnn_states: Dict[str, Tensor],
+    ) -> Tuple[Dict[str, Tensor], Dict[str, Tensor]]:
+        actions = {}
+        next_rnn_states = {}
+        for i, agent in enumerate(self.environment.agents):
+            agent_observation = observations[agent]
+            if self.add_agent_id_to_obs:
+                agent_observation = concat_agent_id_to_obs(
+                    agent_observation, i, len(self.environment.agents)
+                )
+            agent_observation = tf.expand_dims(agent_observation, axis=0)  # add batch dimension
+            action, next_rnn_states[agent] = self.policy_network(
+                agent_observation, rnn_states[agent]
+            )
+
+            # Store agent action
+            actions[agent] = action
+
+        return actions, next_rnn_states
+
+    def train_step(self, experience: Experience) -> Dict[str, Numeric]:
+        logs = self._tf_train_step(experience)
+        return logs  # type: ignore
+
+    @tf.function(jit_compile=True)  # NOTE: comment this out if using debugger
+    def _tf_train_step(self, experience: Dict[str, Any]) -> Dict[str, Numeric]:
+        # Unpack the batch
+        observations = tf.cast(experience["observations"], "float32")  # (B,T,N,O)
+        replay_actions = tf.cast(
+            tf.clip_by_value(experience["actions"], -1.0, 1.0), "float32"
+        )  # (B,T,N,A)
+        env_states = tf.cast(experience["infos"]["state"], "float32")  # (B,T,S)
+        rewards = tf.cast(experience["rewards"], "float32")  # (B,T,N)
+        truncations = tf.cast(experience["truncations"], "float32")  # (B,T,N)
+        terminals = tf.cast(experience["terminals"], "float32")  # (B,T,N)
+
+        # When to reset the RNN hidden state
+        resets = tf.maximum(terminals, truncations)  # equivalent to logical 'or'
+
+        # Get dims
+        B, T, N = replay_actions.shape[:3]
+
+        # Maybe add agent ids to observation
+        if self.add_agent_id_to_obs:
+            observations = batch_concat_agent_id_to_obs(observations)
+
+        # Make time-major
+        observations = switch_two_leading_dims(observations)
+        replay_actions = switch_two_leading_dims(replay_actions)
+        rewards = switch_two_leading_dims(rewards)
+        terminals = switch_two_leading_dims(terminals)
+        env_states = switch_two_leading_dims(env_states)
+        resets = switch_two_leading_dims(resets)
+
+        # Unroll target policy
+        target_actions = unroll_rnn(
+            self.target_policy_network,
+            merge_batch_and_agent_dim_of_time_major_sequence(observations),
+            merge_batch_and_agent_dim_of_time_major_sequence(resets),
+        )
+        target_actions = expand_batch_and_agent_dim_of_time_major_sequence(target_actions, B, N)
+
+        # Target critics
+        target_qs_1 = self.target_critic_network_1(env_states, target_actions)
+        target_qs_2 = self.target_critic_network_2(env_states, target_actions)
+
+        # Take minimum between two target critics
+        target_qs = tf.minimum(target_qs_1, target_qs_2)
+
+        # Compute Bellman targets
+        targets = rewards[:-1] + self.discount * (1 - terminals[:-1]) * tf.squeeze(
+            target_qs[1:], axis=-1
+        )
+
+        # Do forward passes through the networks and calculate the losses
+        with tf.GradientTape(persistent=True) as tape:
+            # Online critics
+            qs_1 = tf.squeeze(
+                self.critic_network_1(env_states, replay_actions),
+                axis=-1,
+            )
+            qs_2 = tf.squeeze(
+                self.critic_network_2(env_states, replay_actions),
+                axis=-1,
+            )
+
+            # Squared TD-error
+            critic_loss_1 = tf.reduce_mean(0.5 * (targets - qs_1[:-1]) ** 2)
+            critic_loss_2 = tf.reduce_mean(0.5 * (targets - qs_2[:-1]) ** 2)
+
+            # Combine critic loss
+            critic_loss = (critic_loss_1 + critic_loss_2) / 2
+
+            ###################
+            ### Policy Loss ###
+            ###################
+            online_actions = unroll_rnn(
+                self.policy_network,
+                merge_batch_and_agent_dim_of_time_major_sequence(observations),
+                merge_batch_and_agent_dim_of_time_major_sequence(resets),
+            )
+            online_actions = expand_batch_and_agent_dim_of_time_major_sequence(online_actions, B, N)
+
+            # Unroll online policy
+            policy_qs_1 = self.critic_network_1(env_states, online_actions)
+            policy_qs_2 = self.critic_network_2(env_states, online_actions)
+            policy_qs = tf.minimum(policy_qs_1, policy_qs_2)
+
+            policy_loss = (
+                -(self.bc_alpha / tf.reduce_mean(tf.abs(policy_qs))) * tf.reduce_mean(policy_qs)
+                + 1e-3 * tf.reduce_mean(tf.square(online_actions))
+                + tf.reduce_mean(tf.square(online_actions - replay_actions))
+            )
+
+        # Train critics
+        variables = (
+            *self.critic_network_1.trainable_variables,
+            *self.critic_network_2.trainable_variables,
+        )
+        gradients = tape.gradient(critic_loss, variables)
+        self.critic_optimizer.apply(gradients, variables)
+
+        # Train policy
+        variables = (*self.policy_network.trainable_variables,)
+        gradients = tape.gradient(policy_loss, variables)
+        self.policy_optimizer.apply(gradients, variables)
+
+        # Update target networks
+        online_variables = (
+            *self.critic_network_1.variables,
+            *self.critic_network_2.variables,
+            *self.policy_network.variables,
+        )
+        target_variables = (
+            *self.target_critic_network_1.variables,
+            *self.target_critic_network_2.variables,
+            *self.target_policy_network.variables,
+        )
+
+        # Soft target update
+        tau = self.target_update_rate
+        for src, dest in zip(online_variables, target_variables):
+            dest.assign(dest * (1.0 - tau) + src * tau)
+
+        del tape
+
+        logs = {
+            "mean_dataset_q_values": tf.reduce_mean((qs_1 + qs_2) / 2),
+            "critic_loss": critic_loss,
+            "policy_loss": policy_loss,
+            "mean_chosen_q_values": tf.reduce_mean((policy_qs_1 + policy_qs_2) / 2),
+        }
+
+        return logs
+
+
+@hydra.main(version_base=None, config_path="configs", config_name="iddpg_bc")
+def run_experiment(cfg: DictConfig) -> None:
+    print(cfg)
+
+    env = get_environment(cfg["task"]["source"], cfg["task"]["env"], cfg["task"]["scenario"], seed=cfg["seed"])
+
+    buffer = FlashbaxReplayBuffer(
+        sequence_length=cfg["replay"]["sequence_length"],
+        sample_period=cfg["replay"]["sample_period"],
+        seed=cfg["seed"],
+    )
+
+    download_and_unzip_vault(cfg["task"]["source"], cfg["task"]["env"], cfg["task"]["scenario"])
+
+    buffer.populate_from_vault(cfg["task"]["source"], cfg["task"]["env"], cfg["task"]["scenario"], cfg["task"]["dataset"])
+
+    wandb_config = {
+        "system": cfg["system_name"],
+        "seed": cfg["seed"],
+        "training_steps": cfg["training_steps"],
+        **cfg["task"],
+        **cfg["replay"],
+        **cfg["system"],
+    }
+    logger = WandbLogger(project=cfg["wandb_project"], config=wandb_config)
+
+    system = IDDPGBCSystem(env, logger, **cfg["system"])
+
+    tf.random.set_seed(cfg["seed"])
+
+    system.train(buffer, training_steps=int(cfg["training_steps"]))
+
+
+if __name__ == "__main__":
+    run_experiment()
