+++ og_marl/tf2_systems/offline/qmix_cql.py
@@ -12,26 +12,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Implementation of MAICQ"""
+"""Implementation of QMIX+CQL"""
 from typing import Any, Dict, Tuple
 
-import copy
 import hydra
-import numpy as np
 from omegaconf import DictConfig
-import sonnet as snt
-import tensorflow as tf
 import tree
-from chex import Numeric
+import numpy as np
+import tensorflow as tf
 from tensorflow import Tensor
+import sonnet as snt
+from chex import Numeric
+import copy
 
 from og_marl.environments import get_environment, BaseEnvironment
 from og_marl.loggers import BaseLogger, WandbLogger
 from og_marl.vault_utils.download_vault import download_and_unzip_vault
 from og_marl.replay_buffers import Experience, FlashbaxReplayBuffer
-from og_marl.tf2_baselines.networks import QMixer
-from og_marl.tf2_baselines.systems.base import BaseOfflineSystem
-from og_marl.tf2_baselines.utils import (
+from og_marl.tf2_systems.networks import QMixer
+from og_marl.tf2_systems.offline.base import BaseOfflineSystem
+from og_marl.tf2_systems.utils import (
     batch_concat_agent_id_to_obs,
     concat_agent_id_to_obs,
     expand_batch_and_agent_dim_of_time_major_sequence,
@@ -45,9 +45,9 @@ from og_marl.tf2_baselines.utils import (
 set_growing_gpu_memory()
 
 
-class MAICQSystem(BaseOfflineSystem):
+class QMIXCQLSystem(BaseOfflineSystem):
 
-    """MAICQ System"""
+    """QMIX+CQL System"""
 
     def __init__(
         self,
@@ -61,8 +61,8 @@ class MAICQSystem(BaseOfflineSystem):
         target_update_period: int = 200,
         learning_rate: float = 3e-4,
         add_agent_id_to_obs: bool = True,
-        icq_advantages_beta: float = 0.1,  # from MAICQ code
-        icq_target_q_taken_beta: int = 1000,  # from MAICQ code
+        num_ood_actions: int = 20,  # CQL
+        cql_weight: float = 3.0,  # CQL
     ):
         super().__init__(
             environment,
@@ -98,36 +98,23 @@ class MAICQSystem(BaseOfflineSystem):
         self.mixer = QMixer(len(self.environment.agents), mixer_embed_dim, mixer_hyper_dim)
         self.target_mixer = QMixer(len(self.environment.agents), mixer_embed_dim, mixer_hyper_dim)
 
-        # ICQ hyper-params
-        self.icq_advantages_beta = icq_advantages_beta
-        self.icq_target_q_taken_beta = icq_target_q_taken_beta
-
-        # Policy Network
-        self.policy_network = snt.DeepRNN(
-            [
-                snt.Linear(linear_layer_dim),
-                tf.nn.relu,
-                snt.GRU(recurrent_layer_dim),
-                tf.nn.relu,
-                snt.Linear(self.environment.num_actions),
-                tf.nn.softmax,
-            ]
-        )
+        self.num_ood_actions = num_ood_actions
+        self.cql_weight = cql_weight
 
     def reset(self) -> None:
-        """Called at the start of a new episode."""
-        # Reset the recurrent neural network
+        """Called at the start of a new episode during evaluation."""
         self.rnn_states = {
-            agent: self.policy_network.initial_state(1) for agent in self.environment.agents
-        }
-        return
+            agent: self.q_network.initial_state(1) for agent in self.environment.agents
+        }  # reset the rnn hidden states
 
     def select_actions(
         self,
         observations: Dict[str, np.ndarray],
         legal_actions: Dict[str, np.ndarray],
     ) -> Dict[str, np.ndarray]:
-        observations = tree.map_structure(tf.convert_to_tensor, observations)
+        observations, legal_actions = tree.map_structure(
+            tf.convert_to_tensor, (observations, legal_actions)
+        )
         actions, next_rnn_states = self._tf_select_actions(
             observations, legal_actions, self.rnn_states
         )
@@ -145,25 +132,22 @@ class MAICQSystem(BaseOfflineSystem):
     ) -> Tuple[Dict[str, Tensor], Dict[str, Tensor]]:
         actions = {}
         next_rnn_states = {}
-        for i, agent in enumerate(self.environment.agents):
+        for i, agent in enumerate(self.agents):
             agent_observation = observations[agent]
-            agent_observation = concat_agent_id_to_obs(
-                agent_observation, i, len(self.environment.agents)
-            )
+            if self.add_agent_id_to_obs:
+                agent_observation = concat_agent_id_to_obs(agent_observation, i, len(self.agents))
             agent_observation = tf.expand_dims(agent_observation, axis=0)  # add batch dimension
-            probs, next_rnn_states[agent] = self.policy_network(
-                agent_observation, rnn_states[agent]
-            )
+            q_values, next_rnn_states[agent] = self.q_network(agent_observation, rnn_states[agent])
 
             agent_legal_actions = legal_actions[agent]
-            masked_probs = tf.where(
-                tf.equal(agent_legal_actions, 1),
-                probs[0],
+            masked_q_values = tf.where(
+                tf.cast(agent_legal_actions, "bool"),
+                q_values[0],
                 -99999999,
             )
+            greedy_action = tf.argmax(masked_q_values)
 
-            # Max Q-value over legal actions
-            actions[agent] = tf.argmax(masked_probs)
+            actions[agent] = greedy_action
 
         return actions, next_rnn_states
 
@@ -173,11 +157,7 @@ class MAICQSystem(BaseOfflineSystem):
         return logs  # type: ignore
 
     @tf.function(jit_compile=True)
-    def _tf_train_step(
-        self,
-        train_step: int,
-        experience: Dict[str, Any],
-    ) -> Dict[str, Numeric]:
+    def _tf_train_step(self, train_step: int, experience: Dict[str, Any]) -> Dict[str, Numeric]:
         # Unpack the batch
         observations = experience["observations"]  # (B,T,N,O)
         actions = experience["actions"]  # (B,T,N)
@@ -212,9 +192,9 @@ class MAICQSystem(BaseOfflineSystem):
         target_qs_out = expand_batch_and_agent_dim_of_time_major_sequence(target_qs_out, B, N)
 
         # Make batch-major again
-        target_q_vals = switch_two_leading_dims(target_qs_out)
+        target_qs_out = switch_two_leading_dims(target_qs_out)
 
-        with tf.GradientTape(persistent=True) as tape:
+        with tf.GradientTape() as tape:
             # Unroll online network
             qs_out = unroll_rnn(self.q_network, observations, resets)
 
@@ -222,74 +202,86 @@ class MAICQSystem(BaseOfflineSystem):
             qs_out = expand_batch_and_agent_dim_of_time_major_sequence(qs_out, B, N)
 
             # Make batch-major again
-            q_vals = switch_two_leading_dims(qs_out)
+            qs_out = switch_two_leading_dims(qs_out)
 
-            # Unroll the policy
-            probs_out = unroll_rnn(self.policy_network, observations, resets)
+            # Pick the Q-Values for the actions taken by each agent
+            chosen_action_qs = gather(qs_out, actions, axis=3, keepdims=False)
 
-            # Expand batch and agent_dim
-            probs_out = expand_batch_and_agent_dim_of_time_major_sequence(probs_out, B, N)
+            # Max over target Q-Values/ Double q learning
+            qs_out_selector = tf.where(
+                tf.cast(legal_actions, "bool"), qs_out, -9999999
+            )  # legal action masking
+            cur_max_actions = tf.argmax(qs_out_selector, axis=3)
+            target_max_qs = gather(target_qs_out, cur_max_actions, axis=-1)
 
-            # Make batch-major again
-            probs_out = switch_two_leading_dims(probs_out)
-
-            # Mask illegal actions
-            probs_out = probs_out * tf.cast(legal_actions, "float32")
-            probs_sum = (
-                tf.reduce_sum(probs_out, axis=-1, keepdims=True) + 1e-10
-            )  # avoid div by zero
-            probs_out = probs_out / probs_sum
-
-            action_values = gather(q_vals, actions)
-            baseline = tf.reduce_sum(probs_out * q_vals, axis=-1)
-            advantages = action_values - baseline
-            advantages = tf.nn.softmax(advantages / self.icq_advantages_beta, axis=0)
-            advantages = tf.stop_gradient(advantages)
-
-            pi_taken = gather(probs_out, actions, keepdims=False)
-            log_pi_taken = tf.math.log(pi_taken)
-
-            coe = self.mixer.k(env_states)
-
-            coma_loss = -tf.reduce_mean(coe * (len(advantages) * advantages * log_pi_taken))
-
-            # Critic learning
-            q_taken = gather(q_vals, actions, axis=-1)
-            target_q_taken = gather(target_q_vals, actions, axis=-1)
-
-            # Mixing critics
-            q_taken = self.mixer(q_taken, env_states)
-            target_q_taken = self.target_mixer(target_q_taken, env_states)
-
-            advantage_Q = tf.nn.softmax(target_q_taken / self.icq_target_q_taken_beta, axis=0)
-            target_q_taken = len(advantage_Q) * advantage_Q * target_q_taken
+            # Q-MIXING
+            chosen_action_qs, target_max_qs, rewards = self.mixing(
+                chosen_action_qs,
+                target_max_qs,
+                env_states,
+                rewards,
+            )
 
             # Compute targets
             targets = (
-                rewards[:, :-1] + (1 - terminals[:, :-1]) * self.discount * target_q_taken[:, 1:]
+                rewards[:, :-1] + (1 - terminals[:, :-1]) * self.discount * target_max_qs[:, 1:]
             )
             targets = tf.stop_gradient(targets)
 
-            # TD error
-            td_error = targets - q_taken[:, :-1]
-            q_loss = 0.5 * tf.square(td_error)
+            # TD-Error Loss
+            td_loss = 0.5 * tf.reduce_mean(tf.square(targets - chosen_action_qs[:, :-1]))
 
-            # Masking
-            q_loss = tf.reduce_mean(q_loss)
+            #############
+            #### CQL ####
+            #############
 
-            # Add losses together
-            loss = q_loss + coma_loss
+            # Sample legal random actions
+            repeated_legals = tf.stack([legal_actions] * self.num_ood_actions, axis=0)
+            repeated_legals = tf.reshape(repeated_legals, (-1, A))
+            random_ood_actions = tf.random.categorical(
+                repeated_legals / tf.reduce_sum(repeated_legals, axis=-1, keepdims=True),
+                1,
+                dtype="int32",
+            )
+            random_ood_actions = tf.reshape(random_ood_actions, (self.num_ood_actions, B, T, N))
 
-        # Apply gradients to policy
+            all_mixed_ood_qs = []
+            for i in range(self.num_ood_actions):
+                # Gather
+                one_hot_indices = tf.one_hot(random_ood_actions[i], depth=qs_out.shape[-1])
+                ood_qs = tf.reduce_sum(
+                    qs_out * one_hot_indices, axis=-1, keepdims=False
+                )  # [B, T, N]
+
+                # Mixing
+                mixed_ood_qs = self.mixer(ood_qs, env_states)  # [B, T, 1]
+                all_mixed_ood_qs.append(mixed_ood_qs)  # [B, T, Ra]
+
+            all_mixed_ood_qs.append(chosen_action_qs)  # [B, T, Ra + 1]
+            all_mixed_ood_qs = tf.concat(all_mixed_ood_qs, axis=-1)
+
+            cql_loss = tf.reduce_mean(
+                tf.reduce_logsumexp(all_mixed_ood_qs, axis=-1, keepdims=True)
+            ) - tf.reduce_mean(chosen_action_qs)
+
+            #############
+            #### end ####
+            #############
+
+            # Add cql_loss to loss
+            loss = td_loss + self.cql_weight * cql_loss
+
+        # Get trainable variables
         variables = (
-            *self.policy_network.trainable_variables,
             *self.q_network.trainable_variables,
             *self.mixer.trainable_variables,
-        )  # Get trainable variables
+        )
 
-        gradients = tape.gradient(loss, variables)  # Compute gradients.
+        # Compute gradients.
+        gradients = tape.gradient(loss, variables)
 
-        self.optimizer.apply(gradients, variables)  # One optimizer for whole system
+        # Apply gradients.
+        self.optimizer.apply(gradients, variables)
 
         # Online variables
         online_variables = (
@@ -309,12 +301,33 @@ class MAICQSystem(BaseOfflineSystem):
                 dest.assign(src)
 
         return {
-            "critic_oss": q_loss,
-            "policy_loss": coma_loss,
+            "loss": loss,
+            "cql_loss": cql_loss,
+            "td_loss": td_loss,
+            "mean_q_values": tf.reduce_mean(qs_out),
+            "mean_chosen_q_values": tf.reduce_mean(chosen_action_qs),
         }
 
+    def mixing(
+        self,
+        chosen_action_qs: Tensor,
+        target_max_qs: Tensor,
+        states: Tensor,
+        rewards: Tensor,
+    ) -> Tuple[Tensor, Tensor, Tensor]:
+        """QMIX"""
+        # VDN
+        # chosen_action_qs = tf.reduce_sum(chosen_action_qs, axis=2, keepdims=True)
+        # target_max_qs = tf.reduce_sum(target_max_qs, axis=2, keepdims=True)
+        # VDN
 
-@hydra.main(version_base=None, config_path="configs", config_name="maicq")
+        chosen_action_qs = self.mixer(chosen_action_qs, states)
+        target_max_qs = self.target_mixer(target_max_qs, states)
+        rewards = tf.reduce_mean(rewards, axis=2, keepdims=True)
+        return chosen_action_qs, target_max_qs, rewards
+
+
+@hydra.main(version_base=None, config_path="configs", config_name="qmix_cql")
 def run_experiment(cfg: DictConfig) -> None:
     print(cfg)
 
@@ -340,7 +353,7 @@ def run_experiment(cfg: DictConfig) -> None:
     }
     logger = WandbLogger(project=cfg["wandb_project"], config=wandb_config)
 
-    system = MAICQSystem(env, logger, **cfg["system"])
+    system = QMIXCQLSystem(env, logger, **cfg["system"])
 
     tf.random.set_seed(cfg["seed"])
