+++ og_marl/tf2_systems/online/configs/maddpg.yaml
@@ -1,0 +1,27 @@
+system_name: maddpg
+seed: 42
+wandb_project: og-marl
+environment_steps: 3e6
+
+task:
+  source: og_marl
+  env: gymnasium_mamujoco
+  scenario: 2reacher 
+
+replay:
+  sequence_length: 10
+  sample_period: 1
+  max_size: 10000
+  store_to_vault: True
+
+system:
+  train_period: 4
+  env_steps_before_train: 5000
+  linear_layer_dim: 64
+  recurrent_layer_dim: 64
+  discount: 0.99
+  target_update_rate: 0.01
+  critic_learning_rate: 1e-3
+  policy_learning_rate: 1e-3
+  add_agent_id_to_obs: True
+  explore_stddev: 0.2
