+++ og_marl/tf2_systems/online/configs/qmix.yaml
@@ -1,0 +1,23 @@
+system_name: qmix
+seed: 42
+wandb_project: og-marl
+environment_steps: 3e6
+
+task:
+  source: og_marl
+  env: smac_v1
+  scenario: 5m_vs_6m 
+
+replay:
+  sequence_length: 20
+  sample_period: 1
+  max_size: 50000
+  store_to_vault: False
+
+system:
+  learning_rate: 3e-4
+  linear_layer_dim: 128
+  recurrent_layer_dim: 128
+  discount: 0.99
+  target_update_period: 200
+  add_agent_id_to_obs: True
